+++ src/utils/error.ts
@@ -15,10 +15,22 @@ export const getMessageFromUnknownError = (error: unknown): string => {
 
 const getErrorMessageFromAxiosError = (error: AxiosError): string => {
   let message: string = 'An unknown network error has occurred.';
-  if (error.response?.status === 401) {
+  if (
+    !error.response &&
+    (error.code === 'ENOTFOUND' ||
+      error.code === 'ECONNREFUSED' ||
+      error.code === 'ENETUNREACH' ||
+      error.message.includes('Network Error'))
+  ) {
+    message = 'No connection could be established. Please check your internet connection and try again.';
+  } else if (error.response?.status === 401) {
     message = 'Your token is no longer valid. Please sign in again.';
   } else if (error.response?.status === 403) {
     message = 'You do not have permission to access this resource.';
+  } else if (error.response?.status === 500) {
+    message = 'An internal server error has occurred. Please try again later.';
+  } else if (error.response?.status === 503) {
+    message = 'The service is currently unavailable. Please try again later.';
   } else if ((error.response?.data as any)?.message) {
     message = (error.response?.data as any)?.message;
   } else if ((error.response?.data as any)?.error?.issues[0]?.message) {